Correctly tweak the correct .config by passing buildpath
We currently default to build-aarch64. This breaks
linux-minimize-defconfig when building a different architecture. Pass
the buildpath from build-kernel.sh so we tweak the right .config.

Similarly, build-env.sh defaults to linux-amit.git as the default kernel
tree. Make cwd as the default kernel tree to allow things to work from
worktrees, for example.

Signed-off-by: Amit Kucheria <amit.kucheria@linaro.org>