build-kernel.sh: Move board control to remote RPi
RPi runs a console server and also controls the power and USB relays. It
is connected to all the boards in the lab through a USB hub.

Add code to check if the console server (RPi) is alive, so we can flash
new images.
Change the code to toggle relays from devbox as before but toggle USB relays from the RPi instead of the devbox.

Signed-off-by: Amit Kucheria <amit.kucheria@linaro.org>